+++ src/__tests__/wildcard-paths.spec.ts
