+++ src/index.ts
@@ -1,0 +1,4 @@
+export { generateSpecs, openAPISpecs } from "./openapi.js";
+export type * from "./types.js";
+export * from "./utils.js";
+export * from "./middlewares.js";
