--- src/old/index.ts
@@ -1,4 +1,0 @@
-export { generateSpecs, openAPISpecs } from "./openapi.js";
-export { describeRoute, generateRouteSpecs } from "./route.js";
-export type * from "./types.js";
-export * from "./utils.js";
